+++ react/AddressContainer.js
@@ -32,7 +32,7 @@ class AddressContainer extends Component {
     }
   }
 
-  handleAddressChange = changedAddressFields => {
+  handleAddressChange = (changedAddressFields) => {
     const {
       cors,
       accountName,
@@ -64,10 +64,10 @@ class AddressContainer extends Component {
     if (
       autoCompletePostalCode &&
       changedAddressFields.postalCode &&
-      !changedAddressFields.postalCode.geolocationAutoCompleted
+      changedAddressFields.postalCode.geolocationAutoCompleted === false
     ) {
       const postalCodeField = rules.fields.find(
-        field => field.name === 'postalCode',
+        (field) => field.name === 'postalCode',
       )
       const diffFromPrev =
         address.postalCode.value !== validatedAddress.postalCode.value