+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- `geolocationAutoCompleted` verification in `handleAddressChange()` to display number input when not written down in geolocation field.
+
 ## [3.10.1] - 2020-05-11
 
 ### Fixed